feat: move border rules from primer/variables to primer/borders
